requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.28.2
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
pydantic==2.7.4
pydantic-core==2.18.4
langchain==0.1.12
langchain-core==0.1.33
langchain-google-genai==0.0.9
langchain-text-splitters==0.0.1
langsmith==0.1.17
elevenlabs==1.54.0 